refactor: slim delivery orchestrator shell [E3.06]
