+++ docs/02-delivery/engineering-epic-03/ticket-06-command-handler-rewire-and-final-orchestrator-slim-down.md
@@ -19,3 +19,13 @@ Finish the modularization by turning the orchestrator file into a composition sh
 - no new commands
 - no redesign of `stacked-closeout`
 - no workflow-semantic expansion
+
+## Rationale
+
+The final extraction moved the remaining top-level orchestration seams out of `tools/delivery/orchestrator.ts` without changing the public `runDeliveryOrchestrator(argv, cwd)` facade.
+
+- `tools/delivery/cli.ts` now owns argv parsing, usage rendering, and plan-path resolution for command execution.
+- `tools/delivery/notifications.ts` now owns notifier resolution, milestone-event mapping, review-window messaging, and best-effort Telegram delivery.
+- `tools/delivery/ticket-flow.ts` now owns ticket progression, handoff rendering, PR-opening orchestration, and restack/advance flow mechanics.
+
+The orchestrator remains the stable facade and runtime composition shell, but no longer directly owns every command concern. User-visible commands, storage roots, PR semantics, and review behavior were intentionally preserved.
